:sparkles: feat: separate server page and client page
