+++ src/components/ui/articles/ArticleProfile/ArticleProfile.tsx
@@ -2,21 +2,31 @@ import type { FC } from 'react';
 
 import Image from 'next/image';
 
-import { Author } from '@/types/article';
+import type { Author } from '@/types/article';
+import type { Color } from '@/types/color';
 
 import DefaultProfile from '~/images/default_profile.jpg';
 
 interface Props {
   author: Author;
+  themeColor: Color;
 }
 
-const ArticleProfile: FC<Props> = ({ author }) => {
+const ArticleProfile: FC<Props> = ({ author, themeColor }) => {
+  const borderColorVariants = {
+    Red: 'border-customRed-500',
+    Orange: 'border-customOrange-500',
+    Green: 'border-customGreen-500',
+    Cyan: 'border-customCyan-500',
+    Violet: 'border-customViolet-500',
+  };
+
   return (
     <div className="flex items-center gap-x-3">
       <Image
         src={author.profileImage?.src || DefaultProfile}
         alt="Profile image"
-        className="size-12 rounded-full border-3 border-hitGray"
+        className={`size-12 rounded-full border-3 ${borderColorVariants[themeColor]}`}
       />
       <div className="flex w-full items-center justify-between">
         <div className="flex flex-col items-start justify-center gap-y-1">
